move docker, k8s, localerr
