+++ internal/cmd/cmd.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"errors"
 	"github.com/airbytehq/abctl/internal/cmd/local"
+	"github.com/airbytehq/abctl/internal/cmd/local/localerr"
 	"github.com/airbytehq/abctl/internal/cmd/version"
-	"github.com/airbytehq/abctl/internal/local/localerr"
 	"github.com/pterm/pterm"
 	"github.com/spf13/cobra"
 	"os"
