+++ internal/cmd/local/k8s/cluster.go
@@ -2,7 +2,7 @@ package k8s
 
 import (
 	"fmt"
-	"github.com/airbytehq/abctl/internal/local/localerr"
+	"github.com/airbytehq/abctl/internal/cmd/local/localerr"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
 	"os"
