+++ internal/local/cmd.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/airbytehq/abctl/internal/local/k8s"
-	"github.com/airbytehq/abctl/internal/local/localerr"
+	k8s2 "github.com/airbytehq/abctl/internal/cmd/local/k8s"
+	"github.com/airbytehq/abctl/internal/cmd/local/localerr"
 	"github.com/airbytehq/abctl/internal/telemetry"
 	"github.com/cli/browser"
 	helmclient "github.com/mittwald/go-helm-client"
@@ -63,11 +63,11 @@ type BrowserLauncher func(url string) error
 
 // Command is the local command, responsible for installing, uninstalling, or other local actions.
 type Command struct {
-	provider k8s.Provider
-	cluster  k8s.Cluster
+	provider k8s2.Provider
+	cluster  k8s2.Cluster
 	http     HTTPClient
 	helm     HelmClient
-	k8s      k8s.K8sClient
+	k8s      k8s2.K8sClient
 	portHTTP int
 	spinner  *pterm.SpinnerPrinter
 	tel      telemetry.Client
@@ -100,7 +100,7 @@ func WithHelmClient(client HelmClient) Option {
 }
 
 // WithK8sClient define the k8s client for this command.
-func WithK8sClient(client k8s.K8sClient) Option {
+func WithK8sClient(client k8s2.K8sClient) Option {
 	return func(c *Command) {
 		c.k8s = client
 	}
@@ -127,7 +127,7 @@ func WithSpinner(spinner *pterm.SpinnerPrinter) Option {
 }
 
 // New creates a new Command
-func New(provider k8s.Provider, portHTTP int, opts ...Option) (*Command, error) {
+func New(provider k8s2.Provider, portHTTP int, opts ...Option) (*Command, error) {
 	c := &Command{provider: provider, portHTTP: portHTTP}
 	for _, opt := range opts {
 		opt(c)
@@ -547,7 +547,7 @@ func ingress() *networkingv1.Ingress {
 }
 
 // defaultK8s returns the default k8s client
-func defaultK8s(kubecfg, kubectx string) (k8s.K8sClient, error) {
+func defaultK8s(kubecfg, kubectx string) (k8s2.K8sClient, error) {
 	k8sCfg, err := k8sClientConfig(kubecfg, kubectx)
 	if err != nil {
 		return nil, fmt.Errorf("%w: %w", localerr.ErrKubernetes, err)
@@ -562,7 +562,7 @@ func defaultK8s(kubecfg, kubectx string) (k8s.K8sClient, error) {
 		return nil, fmt.Errorf("%w: could not create clientset: %w", localerr.ErrKubernetes, err)
 	}
 
-	return &k8s.DefaultK8sClient{ClientSet: k8sClient}, nil
+	return &k8s2.DefaultK8sClient{ClientSet: k8sClient}, nil
 }
 
 // defaultHelm returns the default helm client
